+++ prog_labgen/lab1/__init__.py
@@ -1,4 +1,4 @@
 from .lab1_cli import Lab1CLIParser
-from .lab1_intro import Lab1Task
+from .lab1 import Lab1Task
 
 __all__ = ["Lab1Task", "Lab1CLIParser"]
