+++ prog_labgen/lab1/lab1.py
